Server/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-tenants==3.5.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
celery==5.3.6
redis==5.0.1
python-json-logger==3.3.0
django-storages>=1.13.1
google-cloud-storage>=2.17.0
django-redis==5.4.0
pytest==7.4.4
pytest-django==4.7.0
